refact: split error.h into cpp/h
-
